feat(sprint-3.5): granel, vencimientos, gráficas y sistema de trial
Backend:
- models: Numeric(10,3) en cantidad_actual/DetalleVenta.cantidad, +unidad_medida enum, +fecha_vencimiento
- models: PlanEmpresa enum, +trial_expires_at, +plan en Empresa
- migrations 003 (trial/plan) y 004 (granel/vencimiento) aplicadas en Supabase
- dashboard endpoint: +top_productos (Top 5), +alertas_vencimiento (15 días), +trial info
- schemas: dashboard.py nuevo, producto/venta/empresa actualizados con nuevos campos
- empresas router: POST protegido con x-superadmin-key header
- dependencies: bloqueo automático si trial vencido

Frontend:
- Dashboard: gráfica de barras Recharts (Top 5), panel rojo vencimientos, badge días restantes
- ModalProducto: select unidad_medida + input fecha_vencimiento opcional
- CartSidebar: input decimal para productos granel, subtotal por ítem, subtotal formateado
- CartContext: setCantidad() con validación de stock, manejo granel stock < 1
- Inventario: refresh automático post-venta via custom event

+++ app/schemas/venta.py
@@ -7,11 +7,11 @@ from pydantic import BaseModel, ConfigDict, field_validator, model_validator
 
 class DetalleVentaCrear(BaseModel):
     producto_id: UUID
-    cantidad: int
+    cantidad: Decimal
 
     @field_validator("cantidad")
     @classmethod
-    def cantidad_positiva(cls, v: int) -> int:
+    def cantidad_positiva(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("La cantidad debe ser mayor a 0")
         return v
@@ -38,7 +38,7 @@ class DetalleVentaRespuesta(BaseModel):
 
     id: UUID
     producto_id: UUID
-    cantidad: int
+    cantidad: float
     precio_unitario: float
     subtotal: float
     producto_nombre: str
